Renamed test package to com.nonninz.robomodel to access package private methods
