+++ tests/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.nonninz.robomodel.test"
+    package="com.nonninz.robomodel"
     android:versionCode="1"
     android:versionName="1.0" >
 
@@ -10,7 +10,7 @@
 
     <instrumentation
         android:name="android.test.InstrumentationTestRunner"
-        android:targetPackage="com.nonninz.robomodel.test" />
+        android:targetPackage="com.nonninz.robomodel" />
 
     <application
         android:icon="@drawable/ic_launcher"
